rake: fix for task arguments
This is a rather ugly hack. The aim is to properly work when pressing
Tab after any of the following lines:

    rake quux[foo
    rake quux[foo\ b
    rake quux['foo b
    rake quux['foo bar'
    rake quux['foo bar',
    rake quux['foo bar']

A known bug is that it wil break rather horribly with a more dubious use
of quotes, like:

    rake quux[foo'

For some reason, I observe the following output:

    $ ls -1
    long filename
    $ compgen -o bashdefault -o default -- "'lo"
    long filename

(The expected was for the starting quote to be preserved)

Strangely, this only happens inside the completion, not on a manual
invocation of `compgen`.

Most of the hacks pertain to handling this, and other erratic behavior
regarding quotes.