Drop support for Ruby 2.5 with Rails
Unfortunately, keeping a Rails project with Ruby 2.5 merely afloat is
getting harder and harder. We already had to special-case Nokogiri,
and now Minitest also ceased to be compatible.

Given the limited resources and popularity of this project, the simpler
solution is to drop support for Ruby 2.5 + Rails (5.x, 6.x), and require
a minimum of Ruby 2.6.

Simpler tests that don't involve a full Rails project, like
completion-ruby and completion-gem, still work fine, and therefore still
support Ruby 2.5 and earlier versions, currently going as far back as
Ruby 1.9.